// src/types/ranking-types.ts
export interface Liver {
  rank: number;
  id?: string;
  username: string;
  image_url: string;
  saved_to: string;
  ok: boolean;
  error: null | string;
  user_id: string;
  unique_id: string;
  nickname: string;
  signature: string | null;
  creator_account: string;
  creator_name: string;
  content: string | null;
}